include/uapi/xen/privcmd.h: fix compilation in userspace
xen/interface/xen.h is not exported from kernel headers so remove the
dependency and provide needed defines for domid_t and xen_pfn_t if they
are not already defined by some other e.g. Xen specific headers.

Suggested by Andrew Cooper <andrew.cooper3@citrix.com> on lkml message
<5569F9C9.8000607@citrix.com>.

Use <linux/types.h> based defition for xen_pfn_t so that both uapi and
kernel side headers are compatible.

The ifdef for ARM is ugly but did not find better solutions for it now.

Fixes userspace compilation error:

xen/privcmd.h:38:31: fatal error: xen/interface/xen.h: No such file or directory

+++ arch/x86/include/asm/xen/interface.h
@@ -76,7 +76,7 @@
 /* Explicitly size integers that represent pfns in the public interface
  * with Xen so that on ARM we can have one ABI that works for 32 and 64
  * bit guests. */
-typedef unsigned long xen_pfn_t;
+typedef __u32 xen_pfn_t;
 #define PRI_xen_pfn "lx"
 typedef unsigned long xen_ulong_t;
 #define PRI_xen_ulong "lx"
